requirements.txt
api==0.0.7
config==0.5.0.post0
configparser==4.0.2
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
Jinja2==2.11.0
jinja2-time==0.2.0
jsonify==0.5
jsonpickle==1.2
Keras-Preprocessing==1.1.2
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
pandas==1.3.5
pymongo==3.12.0
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scikit-image==0.14.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sseclient==0.0.27
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tokenizers==0.10.3
torch==1.12.0
torchvision==0.13.0
transformers==4.8.2
unittest2==1.1.0
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==0.16.1
yarl==1.4.2
zipp==3.1.0
